+++ apps/core/serializers/mixins/audit_serializer_mixin.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+
+
+class AuditSerializerMixin(serializers.Serializer):
+    created_by = serializers.HiddenField(
+        default=serializers.CurrentUserDefault()
+    )
+    updated_by = serializers.HiddenField(
+        default=serializers.CurrentUserDefault()
+    )