+++ credit_track/urls.py
@@ -26,6 +26,8 @@ from drf_spectacular.views import (
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/v1/", include("apps.core.urls")),
+    path("api/v1/", include("apps.loans.urls")),
+    path("api/v1/", include("apps.payments.urls")),
     # API and Schema Documentation
     path("api/v1/schema/", SpectacularAPIView.as_view(), name="schema"),
     path(